+++ src/IMAP/Message.php
@@ -269,6 +269,8 @@ class Message {
              * | Mon, 20 Nov 2017 20:31:31 +0800 (GMT+8:00) | Double timezone specification     | A Windows feature
              * |                                            | and invalid timezone (max 6 char) |
              * | 04 Jan 2018 10:12:47 UT                    | Missing letter "C"                | Unknown
+             * | Thu, 31 May 2018 18:15:00 +0800 (added by) | Non-standard details added by the | Unknown
+             * |                                            | mail server                       |
              *
              * Please report any new invalid timestamps to [#45](https://github.com/Webklex/laravel-imap/issues/45)
              */
@@ -276,7 +278,7 @@ class Message {
                 $this->date = Carbon::parse($date);
             } catch (\Exception $e) {
                 switch (true) {
-                    case preg_match('/([A-Z]{2,3}\,\ [0-9]{1,2}\ [A-Z]{2,3}\ [0-9]{4}\ [0-9]{1,2}\:[0-9]{1,2}\:[0-9]{1,2}\ \+[0-9]{4}\ \([A-Z]{2,3}\+[0-9]{1,2}\:[0-9]{1,2})\)+$/i', $date) > 0:
+                    case preg_match('/([A-Z]{2,3}\,\ [0-9]{1,2}\ [A-Z]{2,3}\ [0-9]{4}\ [0-9]{1,2}\:[0-9]{1,2}\:[0-9]{1,2}\ \+[0-9]{4}\ \(.*)\)+$/i', $date) > 0:
                         $array = explode('(', $date);
                         $array = array_reverse($array);
                         $date = trim(array_pop($array));
@@ -589,6 +591,46 @@ class Message {
     }
 
     /**
+     * Find the folder containing this message.
+     *
+     * @param null|Folder $folder where to start searching from (top-level inbox by default)
+     * @return null|Folder
+     */
+    public function getContainingFolder(Folder $folder = null)
+    {
+        $folder = $folder ?: $this->client->getFolders()->first();
+        $this->client->checkConnection();
+
+        // Try finding the message by uid in the current folder
+        $client = new Client;
+        $client->openFolder($folder);
+        $uidMatches = imap_fetch_overview($client->getConnection(), $this->uid, $this->fetch_options);
+        $uidMatch = count($uidMatches)
+            ? new Message($uidMatches[0]->uid, $uidMatches[0]->msgno, $client)
+            : null;
+        $client->disconnect();
+
+        // imap_fetch_overview() on a parent folder will return the matching message
+        // even when the message is in a child folder so we need to recursively
+        // search the children
+        foreach ($folder->children as $child) {
+            $childFolder = $this->getContainingFolder($child);
+
+            if ($childFolder) {
+                return $childFolder;
+            }
+        }
+
+        // before returning the parent
+        if ($this->is($uidMatch)) {
+            return $folder;
+        }
+
+        // or signalling that the message was not found in any folder
+        return null;
+    }
+
+    /**
      * Move the Message into an other Folder
      *
      * @param string  $mailbox
@@ -660,7 +702,7 @@ class Message {
      */
     public function unsetFlag($flag) {
         $flag = "\\".trim(is_array($flag) ? implode(" \\", $flag) : $flag);
-        return imap_clearflag_full($this->client->getConnection(), $this->getUid(), "\\$flag", SE_UID);
+        return imap_clearflag_full($this->client->getConnection(), $this->getUid(), $flag, SE_UID);
     }
 
     /**
@@ -812,5 +854,25 @@ class Message {
      */
     public function getBodies() {
         return $this->bodies;
+    }
+
+    /**
+     * Does this message match another one?
+     *
+     * A match means same uid, message id, subject and date/time.
+     *
+     * @param  null|static $message
+     * @return boolean
+     */
+    public function is(Message $message = null)
+    {
+        if (is_null($message)) {
+            return false;
+        }
+
+        return $this->uid == $message->uid
+            && $this->message_id == $message->message_id
+            && $this->subject == $message->subject
+            && $this->date->eq($message->date);
     }
 }