+++ CHANGELOG.md
@@ -9,10 +9,36 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - NaN
 
 ### Added
+- NaN
+
+### Affected Classes
+- NaN
+
+## [1.1.1] - 2018-05-04
+### Fixed
+- Force to add a space between criteria in search query, otherwise no messages are fetched. Thanks to @cent89
+
+### Added
+- Attachment::getMimeType() and Attachment::getExtension() added
+
+### Affected Classes
+- [Folder::class](src/IMAP/Folder.php)
+- [Attachment::class](src/IMAP/Attachment.php)
+
+## [1.1.0] - 2018-04-24
+### Fixed
+- Client::createFolder($name) fixed #91
+- Versions will now follow basic **Semantic Versioning** guidelines (MAJOR.MINOR.PATCH) 
+
+### Added
 - Connection validation added
+- Client::renameFolder($old_name, $new_name) and Client::deleteFolder($name) methods added #91
+- Find the folder containing a message #92
+- Change all incoming encodings to iconv() supported ones #94
 
 ### Affected Classes
 - [Client::class](src/IMAP/Client.php)
+- [Message::class](src/IMAP/Message.php)
 
 ## [1.0.5.9] - 2018-04-15
 ### Added